requirements.txt
cs50
flask
flask-Session
werkzeug.security
